Move out hierarchy test helpers to a separate file
