+++ tests/test_hierarchy_helpers.hpp
@@ -1,0 +1,224 @@
+/*************************************************************************
+ * Copyright (c) 2017-2018 by the mfmg authors                           *
+ * All rights reserved.                                                  *
+ *                                                                       *
+ * This file is part of the mfmg libary. mfmg is distributed under a BSD *
+ * 3-clause license. For the licensing terms see the LICENSE file in the *
+ * top-level directory                                                   *
+ *                                                                       *
+ * SPDX-License-Identifier: BSD-3-Clause                                 *
+ *************************************************************************/
+
+#ifndef MFMG_TEST_HIERARCHY_HELPERS_HPP
+#define MFMG_TEST_HIERARCHY_HELPERS_HPP
+
+#include "laplace.hpp"
+
+#include <mfmg/dealii_adapters.hpp>
+#include <mfmg/hierarchy.hpp>
+
+#include <deal.II/base/conditional_ostream.h>
+#include <deal.II/dofs/dof_accessor.h>
+#include <deal.II/fe/fe_q.h>
+#include <deal.II/lac/la_parallel_vector.h>
+#include <deal.II/lac/precondition.h>
+#include <deal.II/lac/solver_cg.h>
+#include <deal.II/lac/trilinos_linear_operator.h>
+#include <deal.II/lac/trilinos_vector.h>
+
+#include <boost/property_tree/info_parser.hpp>
+#include <boost/test/data/monomorphic.hpp>
+#include <boost/test/data/test_case.hpp>
+
+#include <random>
+
+namespace bdata = boost::unit_test::data;
+namespace tt = boost::test_tools;
+
+template <int dim>
+class Source : public dealii::Function<dim>
+{
+public:
+  Source() = default;
+
+  virtual double value(dealii::Point<dim> const &,
+                       unsigned int const = 0) const override final
+  {
+    return 0.;
+  }
+};
+
+template <int dim>
+class ConstantMaterialProperty : public dealii::Function<dim>
+{
+public:
+  ConstantMaterialProperty() = default;
+
+  virtual double value(dealii::Point<dim> const &,
+                       unsigned int const = 0) const override final
+  {
+    return 1.;
+  }
+};
+
+template <int dim>
+class LinearXMaterialProperty : public dealii::Function<dim>
+{
+public:
+  LinearXMaterialProperty() = default;
+
+  virtual double value(dealii::Point<dim> const &p,
+                       unsigned int const = 0) const override final
+  {
+    return 1. + p[0];
+  }
+};
+
+template <int dim>
+class LinearMaterialProperty : public dealii::Function<dim>
+{
+public:
+  LinearMaterialProperty() = default;
+
+  virtual double value(dealii::Point<dim> const &p,
+                       unsigned int const = 0) const override final
+  {
+    double value = 1.;
+    for (unsigned int d = 0; d < dim; ++d)
+      value += (1. + d) * p[d];
+
+    return value;
+  }
+};
+
+template <int dim>
+class DiscontinuousMaterialProperty : public dealii::Function<dim>
+{
+public:
+  DiscontinuousMaterialProperty() = default;
+
+  virtual double value(dealii::Point<dim> const &p,
+                       unsigned int const = 0) const override final
+
+  {
+    double value = 10.;
+    for (unsigned int d = 0; d < dim; ++d)
+      if (p[d] > 0.5)
+        value *= value;
+
+    return value;
+  }
+};
+
+template <int dim>
+class MaterialPropertyFactory
+{
+public:
+  static std::shared_ptr<dealii::Function<dim>>
+  create_material_property(std::string const &material_type)
+  {
+    if (material_type == "constant")
+      return std::make_shared<ConstantMaterialProperty<dim>>();
+    else if (material_type == "linear_x")
+      return std::make_shared<LinearXMaterialProperty<dim>>();
+    else if (material_type == "linear")
+      return std::make_shared<LinearMaterialProperty<dim>>();
+    else if (material_type == "discontinuous")
+      return std::make_shared<DiscontinuousMaterialProperty<dim>>();
+    else
+    {
+      mfmg::ASSERT_THROW_NOT_IMPLEMENTED();
+
+      return nullptr;
+    }
+  }
+};
+
+template <int dim, typename VectorType>
+class TestMeshEvaluator : public mfmg::DealIIMeshEvaluator<dim, VectorType>
+{
+private:
+  using value_type =
+      typename mfmg::DealIIMeshEvaluator<dim, VectorType>::value_type;
+
+protected:
+  virtual void evaluate(dealii::DoFHandler<dim> &, dealii::ConstraintMatrix &,
+                        dealii::TrilinosWrappers::SparsityPattern &,
+                        dealii::TrilinosWrappers::SparseMatrix &system_matrix)
+      const override final
+  {
+    // TODO this is pretty expansive, we should use a shared pointer
+    system_matrix.copy_from(_matrix);
+  }
+
+  virtual void
+  evaluate(dealii::DoFHandler<dim> &dof_handler,
+           dealii::ConstraintMatrix &constraints,
+           dealii::SparsityPattern &system_sparsity_pattern,
+           dealii::SparseMatrix<value_type> &system_matrix) const override final
+  {
+    unsigned int const fe_degree = 1;
+    dealii::FE_Q<dim> fe(fe_degree);
+    dof_handler.distribute_dofs(fe);
+
+    dealii::IndexSet locally_relevant_dofs;
+    dealii::DoFTools::extract_locally_relevant_dofs(dof_handler,
+                                                    locally_relevant_dofs);
+
+    // Compute the constraints
+    constraints.clear();
+    constraints.reinit(locally_relevant_dofs);
+    dealii::DoFTools::make_hanging_node_constraints(dof_handler, constraints);
+    constraints.close();
+
+    // Build the system sparsity pattern and reinitialize the system sparse
+    // matrix
+    dealii::DynamicSparsityPattern dsp(dof_handler.n_dofs());
+    dealii::DoFTools::make_sparsity_pattern(dof_handler, dsp, constraints);
+    system_sparsity_pattern.copy_from(dsp);
+    system_matrix.reinit(system_sparsity_pattern);
+
+    // Fill the system matrix
+    dealii::QGauss<dim> const quadrature(fe_degree + 1);
+    dealii::FEValues<dim> fe_values(
+        fe, quadrature,
+        dealii::update_values | dealii::update_gradients |
+            dealii::update_quadrature_points | dealii::update_JxW_values);
+    unsigned int const dofs_per_cell = fe.dofs_per_cell;
+    unsigned int const n_q_points = quadrature.size();
+    dealii::FullMatrix<double> cell_matrix(dofs_per_cell, dofs_per_cell);
+    std::vector<dealii::types::global_dof_index> local_dof_indices(
+        dofs_per_cell);
+    for (auto cell :
+         dealii::filter_iterators(dof_handler.active_cell_iterators(),
+                                  dealii::IteratorFilters::LocallyOwnedCell()))
+    {
+      cell_matrix = 0;
+      fe_values.reinit(cell);
+
+      for (unsigned int q_point = 0; q_point < n_q_points; ++q_point)
+        for (unsigned int i = 0; i < dofs_per_cell; ++i)
+          for (unsigned int j = 0; j < dofs_per_cell; ++j)
+            cell_matrix(i, j) += fe_values.shape_grad(i, q_point) *
+                                 fe_values.shape_grad(j, q_point) *
+                                 fe_values.JxW(q_point);
+
+      cell->get_dof_indices(local_dof_indices);
+      constraints.distribute_local_to_global(cell_matrix, local_dof_indices,
+                                             system_matrix);
+    }
+  }
+
+public:
+  TestMeshEvaluator(dealii::TrilinosWrappers::SparseMatrix const &matrix,
+                    std::shared_ptr<dealii::Function<dim>> material_property)
+      : _matrix(matrix), _material_property(material_property)
+  {
+  }
+
+private:
+  const dealii::TrilinosWrappers::SparseMatrix &_matrix;
+  std::shared_ptr<dealii::Function<dim>> _material_property;
+};
+
+#endif // #ifdef MFMG_TEST_HIERARCHY_HELPERS_HPP
